OCPBUGS-44970: Excluding loki prod-bearer-token container from excessive restarts test in SNO
This container is deployed by https://github.com/openshift/release/tree/master/ci-operator/step-registry/ipi/install/hosted-loki and is sensitive to resolving sso.redhat.com. During an SNO upgrade, there is a brief period of time where dns is unavailable because the only instance of the pod has been killed and the new pod is starting up. The goal of this change is to exclude this container from the excessive restart checks since, depending on the exact timing, it can restart enough times to trigger a failed test.