OCPBUGS-47700: images: Fix test to work with manifestlisted imagestreams
Look at the layers in the submanifest rather than the top level manifest
if the imagestream is manifestlisted. Refer to the bug for details.

This fixes techpreview jobs for clusters installed with the
multi-payload. Example of a failure: https://prow.ci.openshift.org/view/gs/test-platform-results/logs/periodic-ci-openshift-multiarch-master-nightly-4.19-ocp-e2e-aws-ovn-techpreview-multi-a-a/1873687894279852032