Use LB hostname as target if LB IP is not present
AWS exposes Load Balancer services as hostnames instead of as public IP addresses.

Wait for the hostname to resolve to an IP: wait for the DNS record for the corresponding hostname to hold at least one IP address before you go on and check the connection towards the LB service.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>